+++ src/Cdn/exports/Get-AzFrontDoorCdnProfileResourceUsage.ps1
@@ -20,9 +20,7 @@ Checks the quota and actual usage of AzureFrontDoor endpoints under the given CD
 .Description
 Checks the quota and actual usage of AzureFrontDoor endpoints under the given CDN profile.
 .Example
-{{ Add code here }}
-.Example
-{{ Add code here }}
+Get-AzFrontDoorCdnProfileResourceUsage -ResourceGroupName testps-rg-da16jm -ProfileName fdp-v542q6
 
 .Outputs
 Microsoft.Azure.PowerShell.Cmdlets.Cdn.Models.Api20210601.IUsage
@@ -109,7 +107,7 @@ begin {
         $parameterSet = $PSCmdlet.ParameterSetName
 
         if ($null -eq [Microsoft.WindowsAzure.Commands.Utilities.Common.AzurePSCmdlet]::PowerShellVersion) {
-            [Microsoft.WindowsAzure.Commands.Utilities.Common.AzurePSCmdlet]::PowerShellVersion = $Host.Runspace.Version.ToString()
+            [Microsoft.WindowsAzure.Commands.Utilities.Common.AzurePSCmdlet]::PowerShellVersion = $Host.Version.ToString()
         }         
         $preTelemetryId = [Microsoft.WindowsAzure.Commands.Common.MetricHelper]::TelemetryId
         if ($preTelemetryId -eq '') {