+++ src/Sql/Sql/Elastic Jobs/Cmdlet/Agent/AzureSqlElasticJobAgentCmdletBase.cs
@@ -47,7 +47,7 @@ namespace Microsoft.Azure.Commands.Sql.ElasticJobs.Cmdlet
         /// We clear these properties so that during piping scenarios we can ensure we initialize the minimum properties
         /// for either getting, creating, updating, or removing the correct resource
         /// Resource group name, server name, agent name, and name are cleared
-        /// so that during the next iteration in list, they will be initialized properly during <see cref="InitializeInputObjectProperties"/>
+        /// so that during the next iteration in list, they will be initialized properly during <see cref="AzureSqlElasticJobsCmdletBase{TInputObject, TModel, TAdapter}.InitializeInputObjectProperties(TInputObject)"/>
         /// </remarks>
         protected void ClearProperties()
         {