+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,12 @@
 -->
 
 ## Upcoming Release
+* Added support for partition key and id paths to be part of client encryption policy.
+* Fixed bug related to Update-AzCosmosDBSqlContainer command on containers with Client Encryption Policy.
+
+## Version 1.8.1
+* Fixed bug related to Update-AzCosmosDBSqlContainer command on containers with Client Encryption Policy.
+* Fixed the optional Location paramater of New-AzCosmosDBAccount cmdlet.
 
 ## Version 1.8.0
 * Introduced support for creating containers with Client Encryption Policy. The current supported version of Client Encryption Policy is 1.